+++ src/app/app.routes.server.ts
@@ -1,8 +1,17 @@
+// src/app/app.routes.server.ts v3.2.0
 import {RenderMode, ServerRoute} from '@angular/ssr';
 
 export const serverRoutes: ServerRoute[] = [
   {
+    path: '',
+    renderMode: RenderMode.Server,
+  },
+  {
+    path: 'config/:id',
+    renderMode: RenderMode.Server,
+  },
+  {
     path: '**',
-    renderMode: RenderMode.Prerender,
+    renderMode: RenderMode.Server,
   },
 ];